+++ utoipa-swagger-ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog - utoipa-swagger-ui
 
+## Unreleased
+
+### Added
+
+* Allow disabling syntax highlighting (https://github.com/juhaku/utoipa/pull/1188)
+
 ## 8.0.3 - Oct 23 2024
 
 ### Fixed